feat: service 1 & service2, build.sbt refactoring after merging Deps & Dependencies, Config & Helper [K2SO-19]
