+++ project/Config.scala
@@ -1,15 +1,16 @@
+import Dependencies._
 import akka.grpc.sbt.AkkaGrpcPlugin
 import akka.grpc.sbt.AkkaGrpcPlugin.autoImport.akkaGrpcCodeGeneratorSettings
 import com.typesafe.sbt.packager.archetypes.JavaAppPackaging
 import com.typesafe.sbt.packager.docker.DockerPlugin
-import com.typesafe.sbt.packager.docker.DockerPlugin.autoImport.*
+import com.typesafe.sbt.packager.docker.DockerPlugin.autoImport._
 import com.typesafe.sbt.SbtNativePackager.autoImport.maintainer
-import sbt.Keys.*
-import sbt.*
-import scoverage.ScoverageKeys.{coverageFailOnMinimum, *}
+import sbt.Keys._
+import sbt.{Compile, _}
+import scoverage.ScoverageKeys.{coverageFailOnMinimum, _}
 import sbtdynver.DynVerPlugin.autoImport.dynverSeparator
 import sbtdynver.DynVerPlugin.autoImport.dynverVTagPrefix
-import wartremover.WartRemover.autoImport.*
+import wartremover.WartRemover.autoImport._
 
 import scala.collection.immutable.Seq
 import kalix.sbt.KalixPlugin
@@ -26,7 +27,7 @@ import sbtbuildinfo.BuildInfoOption.ToMap
 import sbtbuildinfo.BuildInfoPlugin
 import sbtprotoc.ProtocPlugin.autoImport.PB
 import scalapb.GeneratorOption
-import scalapb.GeneratorOption.{FlatPackage, *}
+import scalapb.GeneratorOption.{FlatPackage, _}
 
 import java.util.Calendar
 
@@ -50,7 +51,7 @@ object Config {
       run / envVars += ("HOST", "0.0.0.0"),
       run / javaOptions ++= Seq("-Dlogback.configurationFile=logback-dev-mode.xml"),
 
-        // NEVER  SET  THIS: version := "0.1"
+      // NEVER  SET  THIS: version := "0.1"
       // IT IS HANDLED BY: sbt-dynver
       ThisBuild / dynverSeparator := "-",
     )
@@ -81,6 +82,7 @@ object Config {
     def withScala2(p: Project): Project = {
       p.configure(withInfo)
         .settings(
+          ThisBuild / dynverSeparator := "-",
           scalaVersion := "2.13.10", // "3.3.1-RC7",
           scalacOptions := scala_2_options,
           apiURL := Some(url("https://riddl.tech/apidoc/")),
@@ -88,7 +90,26 @@ object Config {
           Test / parallelExecution := false,
           Test / testOptions += Tests.Argument("-oDF"),
           Test / logBuffered := false,
-          libraryDependencies ++= Dep.basicTestingDependencies ++ Dep.jsonDependencies
+          libraryDependencies ++= Dependencies.basicTestingDependencies ++ Dependencies.jsonDependencies
+        )
+    }
+
+    def scalapbCodeGen(project: Project): Project = {
+      project.settings(
+        libraryDependencies ++= scalaPbDependencies,
+        Compile / PB.targets := Seq(
+          scalapb.gen(
+            FlatPackage,
+            SingleLineToProtoString,
+            RetainSourceCodeInfo
+          ) -> (Compile / sourceManaged).value / "scalapb",
+          scalapb.validate.gen(
+            FlatPackage,
+            SingleLineToProtoString,
+            RetainSourceCodeInfo
+          ) -> (Compile / sourceManaged).value / "scalapb"
+        ),
+        libraryDependencies += scalaPbCompilerPlugin
       )
     }
 
@@ -108,54 +129,51 @@ object Config {
     }
 
     def withBuildInfo(
-      homePage: String,
-      orgName: String,
-      packageName: String,
-      objName: String = "BuildInfo",
-      baseYear: Int = 2023
+        homePage: String,
+        orgName: String,
+        packageName: String,
+        objName: String = "BuildInfo",
+        baseYear: Int = 2023
     )(p: Project): Project = {
       p.enablePlugins(BuildInfoPlugin)
-      .settings(
-        buildInfoObject := objName,
-        buildInfoPackage := packageName,
-        buildInfoOptions := Seq(ToMap, ToJson, BuildTime),
-        buildInfoUsePackageAsPath := true,
-        buildInfoKeys ++= Seq[BuildInfoKey](
-          name,
-          version,
-          description,
-          organization,
-          organizationName,
-          BuildInfoKey.map(organizationHomepage) { case (k, v) =>
-            k -> v.get.toString
-          },
-          BuildInfoKey.map(homepage) { case (k, v) =>
-            "projectHomepage" -> v.map(_.toString).getOrElse(homePage)
-          },
-          BuildInfoKey.map(startYear) { case (k, v) =>
-            k -> v.map(_.toString).getOrElse(baseYear.toString)
-          },
-          BuildInfoKey.map(startYear) { case (k, v) =>
-            "copyright" -> s"© ${v.map(_.toString).getOrElse(baseYear.toString)}-${
-              Calendar
-                .getInstance()
-                .get(Calendar.YEAR)
-            } $orgName}"
-          },
-          scalaVersion,
-          sbtVersion,
-          BuildInfoKey.map(scalaVersion) { case (k, v) =>
-            val version = if (v.head == '2') {
-              v.substring(0, v.lastIndexOf('.'))
+        .settings(
+          buildInfoObject := objName,
+          buildInfoPackage := packageName,
+          buildInfoOptions := Seq(ToMap, ToJson, BuildTime),
+          buildInfoUsePackageAsPath := true,
+          buildInfoKeys ++= Seq[BuildInfoKey](
+            name,
+            version,
+            description,
+            organization,
+            organizationName,
+            BuildInfoKey.map(organizationHomepage) { case (k, v) =>
+              k -> v.get.toString
+            },
+            BuildInfoKey.map(homepage) { case (k, v) =>
+              "projectHomepage" -> v.map(_.toString).getOrElse(homePage)
+            },
+            BuildInfoKey.map(startYear) { case (k, v) =>
+              k -> v.map(_.toString).getOrElse(baseYear.toString)
+            },
+            BuildInfoKey.map(startYear) { case (k, v) =>
+              "copyright" -> s"© ${v.map(_.toString).getOrElse(baseYear.toString)}-${Calendar
+                  .getInstance()
+                  .get(Calendar.YEAR)} $orgName}"
+            },
+            scalaVersion,
+            sbtVersion,
+            BuildInfoKey.map(scalaVersion) { case (k, v) =>
+              val version = if (v.head == '2') {
+                v.substring(0, v.lastIndexOf('.'))
+              } else v
+              "scalaCompatVersion" -> version
+            },
+            BuildInfoKey.map(licenses) { case (k, v) =>
+              k -> v.map(_._1).mkString(", ")
             }
-            else v
-            "scalaCompatVersion" -> version
-          },
-          BuildInfoKey.map(licenses) { case (k, v) =>
-            k -> v.map(_._1).mkString(", ")
-          }
+          )
         )
-      )
     }
 
     def withWartRemover(proj: Project): Project = {
@@ -170,29 +188,29 @@ object Config {
     }
   }
 
-
   def withDocker(proj: Project): Project = {
     proj
       .enablePlugins(DockerPlugin)
       .settings(
-      dockerBaseImage := "docker.io/library/adoptopenjdk:11-jre-hotspot",
-      dockerUsername := sys.props.get("docker.username"),
-      dockerRepository := sys.props.get("docker.registry"),
-      dockerUpdateLatest := true,
-      dockerBuildCommand := {
-        val arch = sys.props("os.arch")
-        if (arch != "amd64" && !arch.contains("x86")) {
-          // use buildx with platform to build supported amd64 images on other CPU architectures
-          // this may require that you have first run 'docker buildx create' to set docker buildx up
-          dockerExecCommand.value ++ Seq(
-            "buildx",
-            "build",
-            "--platform=linux/amd64",
-            "--load"
-          ) ++ dockerBuildOptions.value :+ "."
-        } else dockerBuildCommand.value
-      }
-    )
+        dockerBaseImage := "docker.io/library/adoptopenjdk:11-jre-hotspot",
+        dockerUsername := sys.props.get("docker.username"),
+        dockerRepository := sys.props.get("docker.registry"),
+        dockerUpdateLatest := true,
+        dockerExposedPorts ++= Seq(8080),
+        dockerBuildCommand := {
+          val arch = sys.props("os.arch")
+          if (arch != "amd64" && !arch.contains("x86")) {
+            // use buildx with platform to build supported amd64 images on other CPU architectures
+            // this may require that you have first run 'docker buildx create' to set docker buildx up
+            dockerExecCommand.value ++ Seq(
+              "buildx",
+              "build",
+              "--platform=linux/amd64",
+              "--load"
+            ) ++ dockerBuildOptions.value :+ "."
+          } else dockerBuildCommand.value
+        }
+      )
   }
 
   object ScalaPB {
@@ -207,15 +225,14 @@ object Config {
       project
         .enablePlugins(AkkaGrpcPlugin)
         .settings(
-          libraryDependencies += Dep.scalaPbValidateCore,
+          libraryDependencies ++= Dependencies.scalaPbValidationDependencies,
           Compile / PB.targets ++= Seq[Target](
             protocbridge.Target(
               scalapb.validate.preprocessor(FlatPackage),
               (Compile / akkaGrpcCodeGeneratorSettings / target).value
             ),
-            protocbridge.Target(
-              scalapb.validate.gen(FlatPackage),
-              (Compile / akkaGrpcCodeGeneratorSettings / target).value)
+            protocbridge
+              .Target(scalapb.validate.gen(FlatPackage), (Compile / akkaGrpcCodeGeneratorSettings / target).value)
           )
         )
     }
@@ -225,12 +242,15 @@ object Config {
 
   object Kalix {
     def service(proj: Project): Project = {
-      proj.enablePlugins(KalixPlugin, JavaAppPackaging, DockerPlugin)
+      proj
+        .enablePlugins(KalixPlugin, JavaAppPackaging, DockerPlugin)
         .configure(Config.Scala.withCoverage(minCoverage))
         .configure(Config.Scala.withScala2)
         .configure(Config.ScalaPB.protoGenValidate)
         .configure(Config.withDocker)
         .settings(
+          exportJars := true,
+          run / envVars += ("HOST", "0.0.0.0"),
           // needed for the proxy to access the user function on all platforms
           run / javaOptions ++= Seq(
             "-Dkalix.user-function-interface=0.0.0.0",
@@ -238,28 +258,52 @@ object Config {
           ),
           run / fork := true,
           Global / cancelable := false, // ctrl-c
-          libraryDependencies ++= Dep.testing ++ Dep.grpc ++ Dep.loggingDependencies
+          libraryDependencies ++= Dependencies.testingDeps ++ Dependencies.grpc ++ Dependencies.loggingDependencies,
+          Compile / scalacOptions ++= Seq(
+            "-target:11",
+            "-deprecation",
+            "-feature",
+            "-unchecked",
+            "-Xlog-reflective-calls",
+            "-Xlint"
+          ),
+          Compile / javacOptions ++= Seq(
+            "-Xlint:unchecked",
+            "-Xlint:deprecation",
+            "-parameters" // for Jackson
+          )
         )
     }
 
     def library(proj: Project): Project = {
-      proj.enablePlugins(KalixPlugin, JavaAppPackaging)
+      proj
+        .enablePlugins(KalixPlugin, JavaAppPackaging)
         .configure(Config.Scala.withScala2)
         .configure(Config.Scala.withCoverage(minCoverage))
         .configure(Config.ScalaPB.protoGenValidate)
         .settings(
-          libraryDependencies ++= Dep.testing
+          libraryDependencies ++= testingDeps
         )
     }
 
     def dependsOn(dependency: Project)(
-      project: Project
+        project: Project
     ): Project = {
       project
         .dependsOn(dependency)
-        .dependsOn( dependency % "protobuf")
+        .dependsOn(dependency % "protobuf;compile->compile;test->test")
     }
 
   }
 }
 
+object Testing {
+  def scalaTest(project: Project): Project = {
+    project.settings(
+      Test / parallelExecution := false,
+      Test / testOptions += Tests.Argument("-oDF"),
+      Test / logBuffered := false,
+      libraryDependencies ++= basicTestingDependencies ++ jsonDependencies
+    )
+  }
+}
