+++ service1/src/main/proto/com/improving/template/domain/service1Commands.proto
@@ -1,0 +1,9 @@
+syntax = "proto3";
+
+package com.improving.template.domain;
+
+import "com/improving/template/domain/service1DomainObjects.proto";
+
+message DoNothingCommand {
+    DoNothing nothing = 1;
+}
