+++ service2/src/main/proto/com/improving/template/domain/service2DomainObjects.proto
@@ -1,0 +1,7 @@
+syntax = "proto3";
+
+package com.improving.template.domain;
+
+message DoNothing {}
+
+message NoState {}
